MySQL: Dereference scalar values in extra fields
The parser will always generate a scalar reference for CURRENT_TIMESTAMP
(to avoid quoting), but it only dereferences it for DEFAULT, not for ON
UPDATE, causing producers to emit SCALAR(0x........) instead.

So here we just dereference scalar references when emitting values for
extra fields. It's only for MySQL and won't help other producers at all.